Updated SimManager for dirty bit control and database functionality
Likely need a mongoDB local server running (or to configure SimManager scripts to run without a database connection.

Currently experiences runtime errors, unsure if this was due to a broken reference in NavMesh or Actor generation. Once runtime is fixed and validity testing is performed, branch should be merged and closed